browser/components/payments/test/mochitest/formautofill/test_editCreditCard.html

<!DOCTYPE HTML>
<html>
<!--
Test that editCreditCard.xhtml is accessible for tests in the parent directory.
-->
<head>
  <meta charset="utf-8">
  <title>Test that editCreditCard.xhtml is accessible</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/AddTask.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <p id="display">
    <iframe id="editCreditCard" src="editCreditCard.xhtml"></iframe>
  </p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
<script type="application/javascript">

add_task(async function test_editCreditCard() {
  let editCreditCard = document.getElementById("editCreditCard").contentWindow;
  await SimpleTest.promiseFocus(editCreditCard);
  ok(editCreditCard.document.getElementById("form"), "Check form is present");
  ok(editCreditCard.document.getElementById("cc-number"), "Check cc-number is present");
});

</script>

</body>
</html>
